refactor: move stray components to root
