+++ app/blog/page.tsx
@@ -1,5 +1,5 @@
 import { Suspense } from 'react'
-import BlogClient from './BlogClient'
+import BlogClient from '@/components/organisms/BlogClient'
 import LoadingOverlay from '@/components/organisms/LoadingOverlay'
 import type { Metadata } from 'next'
 import { getPostsFromPB } from '@/lib/posts/getPostsFromPB'
