+++ app/loja/produtos/[slug]/page.tsx
@@ -4,7 +4,7 @@ import { useParams } from 'next/navigation'
 import Link from 'next/link'
 import { Suspense } from 'react'
 import LoadingOverlay from '@/components/organisms/LoadingOverlay'
-import ProdutoInterativo from './ProdutoInterativo'
+import ProdutoInterativo from '@/components/organisms/ProdutoInterativo'
 import type { Produto as ProdutoBase } from '@/types'
 export default function ProdutoDetalhe() {
   const { slug } = useParams<{ slug: string }>()
