+++ components/molecules/index.ts
@@ -3,3 +3,6 @@ export { InputWithMask } from './InputWithMask'
 export { default as SmoothTabs } from './SmoothTabs'
 export { default as SaldoCard } from './SaldoCard'
 export { default as NextPostButton } from './NextPostButton'
+export { default as CartButton } from './CartButton'
+export { default as CartPreview } from './CartPreview'
+export { default as AddToCartButton } from './AddToCartButton'
