+++ components/organisms/index.ts
@@ -4,3 +4,9 @@ export { default as BlogHeroCarousel } from './BlogHeroCarousel'
 export { default as BlogSidebar } from './BlogSidebar'
 export { default as PostSuggestions } from './PostSuggestions'
 export { default as InscricaoForm } from './InscricaoForm'
+export { default as AuthModal } from './AuthModal'
+export { default as BlogClient } from './BlogClient'
+export { default as Hero } from './Hero'
+export { default as ProdutoInterativo } from './ProdutoInterativo'
+export { default as ProdutosFiltrados } from './ProdutosFiltrados'
+export { default as ProdutosFiltradosCategoria } from './ProdutosFiltradosCategoria'
