+++ components/templates/index.ts
@@ -5,3 +5,5 @@ export { default as HeaderAdmin } from './HeaderAdmin'
 export { default as LayoutWrapperAdmin } from './LayoutWrapperAdmin'
 export { default as HeaderLoja } from './HeaderLoja'
 export { default as FooterLoja } from './FooterLoja'
+export { default as LoginForm } from './LoginForm'
+export { default as SignUpForm } from './SignUpForm'
