+++ stories/AuthModal.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/nextjs'
-import AuthModal from '../app/components/AuthModal'
+import AuthModal from '@/components/organisms/AuthModal'
 import { AuthProvider } from '../lib/context/AuthContext'
 import { ThemeProvider } from '../lib/context/ThemeContext'
